+++ client/src/api/api.tsx
@@ -1,6 +1,5 @@
 import axios, { AxiosResponse, AxiosError } from "axios";
 import { PokemonStats, Pokemon } from "./ApiInterface";
-
 export async function fetchPokemonStats(): Promise<any> {
   try {
     // Make the API call using Axios
@@ -9,11 +8,7 @@ export async function fetchPokemonStats(): Promise<any> {
     );
 
     // Extract the data from the response
-    const results = {
-      species: response.data.species,
-      stats: response.data.stats,
-    };
-    return results;
+    return response.data;
   } catch (error) {
     // Handle errors
     if (axios.isAxiosError(error)) {
@@ -33,36 +28,13 @@ export async function fetchPokemonStats(): Promise<any> {
   }
 }
 
-export async function fetchPokemonStatsProcess(): Promise<PokemonStats> {
-  const data: any = await fetchPokemonStats();
-  let pokemonStats: PokemonStats = {};
-  pokemonStats.name = data.species.name;
-  data.stats.map((element: any) => {
-    if (element.stat.name == "hp") {
-      pokemonStats.hp = element.base_stat;
-    } else if (element.stat.name == "attack") {
-      pokemonStats.attack = element.base_stat;
-    } else if (element.stat.name == "defence") {
-      pokemonStats.defence = element.base_stat;
-    } else if (element.stat.name == "special-attack") {
-      pokemonStats.specialAttack = element.base_stat;
-    } else if (element.stat.name == "special-defence") {
-      pokemonStats.specialDefence = element.base_stat;
-    } else if (element.stat.name == "speed") {
-      pokemonStats.speed = element.base_stat;
-    }
-  });
-
-  return pokemonStats;
-}
-
-export async function fetchPokedexDataByGeneration(): Promise<any> {
+export async function fetchPokedexByGeneration(): Promise<any> {
   try {
     // Make the API call using Axios
     const response: AxiosResponse = await axios.get(
       "https://pokeapi.co/api/v2/pokedex/3"
     );
-    return response.data.pokemon_entries;
+    return response.data;
   } catch (error) {
     // Handle errors
     if (axios.isAxiosError(error)) {
@@ -80,17 +52,4 @@ export async function fetchPokedexDataByGeneration(): Promise<any> {
     }
     throw error;
   }
-}
-
-export async function fetchPokedexDataByGenerationProcess(): Promise<Pokemon[]> {
-  const data: any = await fetchPokedexDataByGeneration();
-  let pokedex: Pokemon[] = [];
-  data.map((element: any) => {
-    let pokemon: Pokemon = {};
-    pokemon.entry = element.entry_number;
-    pokemon.name = element.pokemon_species.name;
-    pokemon.generation = 1;
-    pokedex.push(pokemon)
-  });
-  return pokedex;
 }
